% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfdMethods.R
\name{$<-.nfd}
\alias{$<-.nfd}
\title{Title}
\usage{
\method{$}{nfd}(x, name) <- value
}
\arguments{
\item{x}{An object of class `nfd`.}

\item{name}{The name of the element to access.}

\item{value}{The value of the element to assign}
}
\value{
assign value to requested element.
}
\description{
Title
}
\examples{
mat <- matrix(1:9, nrow = 3)
nfd_obj <- nfd(mat)
nfd_obj$spars_par <- c(1,2)
}
