% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{-.hd}
\alias{-.hd}
\title{Subtraction of two `hd` objects}
\usage{
\method{-}{hd}(obj1, obj2 = NULL)
}
\arguments{
\item{obj1}{An `hd` object}

\item{obj2}{An optional `hd` object}
}
\value{
An `hd` object
}
\description{
Subtraction of two `hd` objects
}
\seealso{
\code{\link{hd}},\code{\link{mvbasismfd}}
}
