% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{functional_ICC}
\alias{functional_ICC}
\title{Functional Intraclass Correlation}
\usage{
functional_ICC(MHPCA)
}
\arguments{
\item{MHPCA}{MHPCA object}
}
\value{
fICC (data.frame)
}
\description{
Functional Intraclass Correlation
}
