% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{functional_ICC_bootstrap}
\alias{functional_ICC_bootstrap}
\title{Functional Intraclass Correlation Bootstrap}
\usage{
functional_ICC_bootstrap(
  MHPCA,
  nboots,
  alpha = 0.05,
  maxiter,
  epsilon,
  fve_cutoff,
  nknots,
  quiet = FALSE,
  file_name = NULL
)
}
\arguments{
\item{MHPCA}{MHPCA object}

\item{nboots}{number of bootstraps to perform}

\item{alpha}{alpha level}

\item{maxiter}{maximum iterations}

\item{epsilon}{epsilon for MM algorithm}

\item{fve_cutoff}{fraction of variance cutoff for M-HPCA}

\item{nknots}{number of knots for splines}

\item{quiet}{logical for printing timing messages}

\item{file_name}{name of file to save individual bootstrapped results in if desired}
}
\value{
fICC percentiles (data.frame)
}
\description{
Functional Intraclass Correlation Bootstrap
}
