% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdClass.R
\name{hd}
\alias{hd}
\alias{Hd}
\title{Define a Set of Multivariate Multidimensional Functional Data objects}
\usage{
Hd(...)
}
\arguments{
\item{...}{A `mfd` objects which have separated by comma}
}
\description{
The `hd` class represents functional data ...

Constructor for `hd` objects (same as `Hd`)
}
\examples{
require(fda)
bs1 <- create.fourier.basis(c(0, 2 * pi), 5)
bs2 <- create.bspline.basis(c(0, 1), 7)
bs3 <- create.exponential.basis(c(0, 2), 3)
nobs <- 10
argval1 <- seq(0, 2 * pi, length.out = 12)
X1 <- outer(sin(argval1), seq(0.5, 1.5, length.out = nobs))
mdbs1 <- Basismfd(bs1)
mfd1 <- Mfd(argval1, X1, mdbs1)
mdbs2 <- Basismfd(bs1)
argval2 <- argval1
X2 <- outer(cos(argval2), seq(0.2, 1.5, length.out = nobs))
mfd2 <- Mfd(argval2, X2, mdbs1)
hd1 <- Hd(mfd1, mfd2)
hd1[1]
hd1[1, 1]
hd1[1:5, 2]
hd1[, 1]
hd1[1:5, ]
evalarg <- list(argval1, argval2)
hd1$eval(evalarg)
hd1 + hd1
mean(hd1)
inprod_hd(hd1, hd1)
norm_hd(hd1)
plot(hd1)
bimfdplot(hd1)

}
\seealso{
\code{\link{mvbasismfd}}, \code{\link{mfd}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{basis}}{A `mvbasismfd` object}

\item{\code{coefs}}{a matrix of the coefficients.}

\item{\code{nobs}}{number of observation}

\item{\code{nvar}}{number of variables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hd-new}{\code{hd$new()}}
\item \href{#method-hd-eval}{\code{hd$eval()}}
\item \href{#method-hd-print}{\code{hd$print()}}
\item \href{#method-hd-clone}{\code{hd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hd-new"></a>}}
\if{latex}{\out{\hypertarget{method-hd-new}{}}}
\subsection{Method \code{new()}}{
Constructor for `hd` objects (same as 'Hd')
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hd$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A `mfd` objects which have separated by comma}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hd-eval"></a>}}
\if{latex}{\out{\hypertarget{method-hd-eval}{}}}
\subsection{Method \code{eval()}}{
Eval method for `hd` objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hd$eval(evalarg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{evalarg}}{A list of numeric vectors of argument values at which the `hd` is to be evaluated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of evaluated values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hd-print"></a>}}
\if{latex}{\out{\hypertarget{method-hd-print}{}}}
\subsection{Method \code{print()}}{
Print method for `hd` objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hd$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to `print`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
