% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_pca.R
\name{hybrid_pca}
\alias{hybrid_pca}
\title{Factory function to create either mhpca or remfpca objects}
\usage{
hybrid_pca(
  hd_obj,
  method = "power",
  ncomp = 3,
  smooth_tuning = NULL,
  sparse_tuning = NULL,
  centerfns = TRUE,
  alpha_orth = FALSE,
  smoothing_type = "coefpen",
  sparse_type = "soft",
  K_fold = 30,
  sparse_CV = TRUE,
  smooth_GCV = TRUE,
  penalize_nfd = FALSE,
  penalize_u = FALSE
)
}
\arguments{
\item{hd_obj}{}

\item{method}{}

\item{ncomp}{}

\item{smooth_tuning}{}

\item{sparse_tuning}{}

\item{centerfns}{}

\item{alpha_orth}{}

\item{smoothing_type}{}

\item{sparse_type}{}

\item{K_fold}{}

\item{sparse_CV}{}

\item{smooth_GCV}{}

\item{penalize_nfd}{}

\item{penalize_u}{}
}
\description{
Factory function to create either mhpca or remfpca objects
}
