% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{inprod_hd}
\alias{inprod_hd}
\title{Compute the inner product between two objects of class `hd`}
\usage{
inprod_hd(hd_obj1, hd_obj2)
}
\arguments{
\item{hd_obj1}{An `hd` object}

\item{hd_obj2}{An `hd` object}
}
\value{
The inner products matrix between the two `hd` objects
}
\description{
Compute the inner product between two objects of class `hd`
}
\seealso{
\code{\link{hd}},\code{\link{mvbasismfd}}
}
