% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnfdMethods.R
\name{inprod_mvnfd}
\alias{inprod_mvnfd}
\title{Compute the inner product between two objects of class `mvnfd`}
\usage{
inprod_mvnfd(mvnfd_obj1, mvnfd_obj2)
}
\arguments{
\item{mvnfd_obj1}{An `mvnfd` object}

\item{mvnfd_obj2}{An `mvnfd` object}
}
\value{
The inner products matrix between the two `mvnfd` objects
}
\description{
Compute the inner product between two objects of class `mvnfd`
}
\seealso{
\code{\link{mvnfd}}
}
