% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfdMethods.R
\name{inprod_nfd}
\alias{inprod_nfd}
\title{Compute the inner product between two objects of class `nfd`}
\usage{
inprod_nfd(nfd_obj1, nfd_obj2)
}
\arguments{
\item{nfd_obj1}{An `nfd` object}

\item{nfd_obj2}{An `nfd` object}
}
\value{
The inner products matrix between the two `nfd` objects
}
\description{
Compute the inner product between two objects of class `nfd`
}
\seealso{
\code{\link{nfd}}
}
