% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.hd}
\alias{is.hd}
\title{Check if an object is of class 'hd'}
\usage{
is.hd(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
TRUE if the object is of class 'hd', FALSE otherwise.
}
\description{
Check if an object is of class 'hd'
}
\seealso{
\code{\link{is.hd}}, \code{\link{is.mfd}}, \code{\link{is.basismfd}}
}
