% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mvnfd}
\alias{is.mvnfd}
\title{Check if an object is of class 'mvnfd'}
\usage{
is.mvnfd(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
TRUE if the object is of class 'mvnfd', FALSE otherwise.
}
\description{
Check if an object is of class 'mvnfd'
}
\seealso{
\code{\link{is.nfd}}, \code{\link{is.mfd}}, \code{\link{is.basismfd}}
}
