% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.nfd}
\alias{is.nfd}
\title{Check if an object is of class 'nfd'}
\usage{
is.nfd(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
TRUE if the object is of class 'nfd', FALSE otherwise.
}
\description{
Check if an object is of class 'nfd'
}
\seealso{
\code{\link{is.nfd}}, \code{\link{is.mfd}}, \code{\link{is.basismfd}}
}
