% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.vd}
\alias{is.vd}
\title{Check if an object is of class 'vd'}
\usage{
is.vd(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
TRUE if the object is of class 'vd', FALSE otherwise.
}
\description{
Check if an object is of class 'vd'
}
\seealso{
\code{\link{is.vd}}, \code{\link{is.mfd}}, \code{\link{is.basismfd}}
}
