% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-alg.R
\name{loglik_calc_complete}
\alias{loglik_calc_complete}
\title{Log-likelihood calculation for MM algorithm with complete data}
\usage{
loglik_calc_complete(
  sigma2,
  TRJ,
  n,
  z_i,
  w_i,
  E_sqrt,
  F_mat,
  F_sqrt,
  I_G,
  I_HJ,
  omegainv_y,
  yt_omegainv_y,
  wtw,
  ztz,
  wtz
)
}
\arguments{
\item{sigma2}{sigma2}

\item{TRJ}{TRJ}

\item{n}{n}

\item{z_i}{z_i}

\item{w_i}{w_i}

\item{E_sqrt}{E_sqrt}

\item{F_mat}{F_mat}

\item{F_sqrt}{F_sqrt}

\item{I_G}{I_G}

\item{I_HJ}{I_HJ}

\item{omegainv_y}{omegainv_y}

\item{yt_omegainv_y}{yt_omegainv_y}

\item{wtw}{wtw}

\item{ztz}{ztz}

\item{wtz}{wtz}
}
\description{
Log-likelihood calculation for MM algorithm with complete data
}
