% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-tests.R
\name{MHPCA_bootstrap_between}
\alias{MHPCA_bootstrap_between}
\title{Multilevel-Hybrid Principal Component Analysis Bootstrap Test}
\usage{
MHPCA_bootstrap_between(MHPCA, B, region, group, nknots, quiet = FALSE)
}
\arguments{
\item{MHPCA}{MHPCA output from MHPCA function}

\item{B}{number of bootstrap samples (scalar)}

\item{region}{region to consider for hypothesis test (scalar)}

\item{group}{groups to consider for the hypothesis test (character)}

\item{nknots}{number of knots for smoothing splines}

\item{quiet}{print timing messages (logical)}
}
\description{
Function for performing the parametric bootstrap test for
comparing groups as laid out in 'Multilevel-Hybrid Principal Component
Analysis' by Campos et. al (20??).
}
