% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-hpca.R
\name{MHPCA_decomp}
\alias{MHPCA_decomp}
\title{Multilevel Hybrid Principal Component Analysis}
\usage{
MHPCA_decomp(
  data,
  fve_cutoff,
  nknots,
  maxiter = 1000,
  epsilon = 0.001,
  reduce = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data}{dataframe in long format with six labeled columns
(Repetition: (character vector),
Subject: subject IDs (character vector),
Group: subject group (character vector),
func: functional argument (numeric vector),
reg: regional argument (character vector),
y: region-referenced functional data (numeric vector))
and row length equal to the length of the vectorized region-referenced
Repetitions across all subjects and groups}

\item{fve_cutoff}{fraction of variance cutoff for reducing the number of product components used in the mixed effects model (scalar in (0, 1))}

\item{nknots}{number of knots to use for smoothing splines}

\item{maxiter}{maximum number of iterations for MM algorithm (scalar)}

\item{epsilon}{epsilon value for determining log-likelihood convergence (scalar)}

\item{reduce}{should the number of product components be reduced and the mixed effects model re-estimated (logical)}

\item{quiet}{display messages for timing (logical)}
}
\value{
A list with
\itemize{
\item mu: the overall mean function (vector),
\item eta: group-region-level-specific shifts (dataframe),
\item covar: list with total, between and within covariance matrices,
\item marg: list with between and within marginal covariances,
\item model: list of models for each group, including scores and variances,
\item data: data with all of the different pieces from the estimation,
\item FVE: fraction of variance explained.
}
}
\description{
Function for performing MHPCA decomposition described in
"Multilevel Hybrid Principal Components Analysis For Region-Referenced
Functional EEG Data" by Campos et al. (202?), including estimation of
fixed effects and marginal covariance functions, marginal eigencompoents,
subject-specific scores, variance components, and measurement error
variance.
}
