% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{MHPCA_simulation}
\alias{MHPCA_simulation}
\title{M-HPCA Simulation}
\usage{
MHPCA_simulation(
  sig_eps,
  n_d,
  J,
  D,
  num_reg,
  num_time,
  missing_level = FALSE,
  K = 2,
  L = 2,
  P = 2,
  M = 2
)
}
\arguments{
\item{sig_eps}{error standard deviation (scalar)}

\item{n_d}{group sample size (scalar)}

\item{J}{number of repetitions per subject (scalar)}

\item{D}{number of groups (scalar)}

\item{num_reg}{number of regions (scalar)}

\item{num_time}{number of functional points (scalar)}

\item{missing_level}{allow for missing repetitions (logical)}

\item{K}{number of marginal level 1 eigenvectors (scalar)}

\item{L}{number of marginal level 1 eigenfunctions (scalar)}

\item{P}{number of marginal level 2 eigenvectors (scalar)}

\item{M}{number of marginal level 2 eigenfunctions (scalar)}
}
\value{
A list with the simulated components.
\itemize{
\item mu - The simulated mean function
\item eta - The simulated group-region-repetition shift from the mean
\item v_k - The simulated level 1 eigenvectors
\item phi_l - the simulated level 1 eigenfunctions
\item lambda_kl - The simulated level 1 eigenvalues
\item v_p - The simulated level 2 eigenvectors
\item phi_m - The simulated level 2 eigenfunctions
\item lambda_pm - The simulated level 2 eigenvalues
\item xi - The simulated subject-specific scores
\item zeta - The simulated subject/repetition-specific scores
\item data - The simulated dataframe
\item sig_eps - The simulated measurement error variance
}
}
\description{
Function for simulating data as described in the supplementary
material of Campos et al. (202?)
}
