% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{MHPCA_simulation_within_group_test}
\alias{MHPCA_simulation_within_group_test}
\title{MHPCA Simulation for Group-Level Inference}
\usage{
MHPCA_simulation_within_group_test(
  sig_eps,
  n_d,
  J,
  D,
  num_reg,
  num_time,
  missing_level = FALSE,
  test_group,
  test_region,
  delta
)
}
\arguments{
\item{sig_eps}{error standard deviation (scalar)}

\item{n_d}{group sample size (scalar)}

\item{J}{number of repetitions per subject (scalar)}

\item{D}{number of groups (scalar)}

\item{num_reg}{number of regions (scalar)}

\item{num_time}{number of functional points (scalar)}

\item{missing_level}{allow for missing repetitions (logical)}

\item{test_group}{test group (scalar)}

\item{test_region}{test region (scalar)}

\item{delta}{tuning parameter (scalar)}
}
\value{
A list with the simulated components.
\itemize{
\item mu - The simulated mean function
\item eta - The simulated group-region-repetition shift from the mean
\item v_k - The simulated level 1 eigenvectors
\item phi_l - the simulated level 1 eigenfunctions
\item lambda_kl - The simulated level 1 eigenvalues
\item v_p - The simulated level 2 eigenvectors
\item phi_m - The simulated level 2 eigenfunctions
\item lambda_pm - The simulated level 2 eigenvalues
\item xi - The simulated subject-specific scores
\item zeta - The simulated subject/repetition-specific scores
\item data - The simulated dataframe with columns for
\item sig_eps - The simulated measurement error variance
}
}
\description{
Function for simulating data for group-level inference as
described in the supplementary materials.
}
