% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-alg.R
\name{mm}
\alias{mm}
\title{MM Algorithm for M-HPCA}
\usage{
mm(MHPCA, group, maxiter = 1000, epsilon = 1e-06, j_tot)
}
\arguments{
\item{MHPCA}{MHPCA object}

\item{group}{group}

\item{maxiter}{maximum iterations}

\item{epsilon}{epsilon value}

\item{j_tot}{total repetitions}
}
\description{
MM Algorithm for M-HPCA
}
