% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm-alg.R
\name{mm_complete}
\alias{mm_complete}
\title{MM Algorithm for M-HPCA with complete data}
\usage{
mm_complete(MHPCA, group, maxiter, epsilon)
}
\arguments{
\item{MHPCA}{MHPCA object}

\item{group}{group}

\item{maxiter}{maximum iterations}

\item{epsilon}{epsilon value}
}
\description{
MM Algorithm for M-HPCA with complete data
}
