% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotionSenseDataset.R
\name{MotionSenseDataset}
\alias{MotionSenseDataset}
\title{Motion Sense Dataset: Measurements of user acceleration and pitch attitude collected by smartphones from 24 individuals performing four distinct activities: jogging, walking, sitting, and standing.}
\format{
A functional time series object of class `funts` with the following fields:
}
\description{
This dataset includes time-series data generated by accelerometer and gyroscope sensors.
A total of 24 participants in a range of gender, age, weight, and height performed 4 activities in the same environment 
and conditions: downstairs, upstairs, walking, jogging, sitting, and standing.
}
\examples{
\dontrun{
# Load the Motion Sense Dataset
data("motion_sense_data")
}
}
