% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvimgClass.R
\name{Mvimg}
\alias{Mvimg}
\title{A Class of Multivariate Image Data Objects}
\usage{
Mvimg(...)
}
\arguments{
\item{...}{Multiple `img` objects to include in the multivariate image set.}
}
\description{
Constructor for `mvimg` objects (same as Mvimg(...))
}
