% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnfdClass.R
\name{mvnfd}
\alias{mvnfd}
\alias{Mvnfd}
\title{Define a Set of Multivariate Multidimensional Vector Data objects}
\usage{
Mvnfd(...)
}
\arguments{
\item{...}{A `nfd` objects which have separated by comma}
}
\description{
The `mvnfd` class represents a set of multivariate multidimensional vector data.

Constructor for `mvnfd` objects (same as Mvnfd(...))
}
\examples{
# Create some vector data
data_matrix1 <- matrix(rnorm(100), nrow = 10, ncol = 10)
data_matrix2 <- matrix(rnorm(100), nrow = 10, ncol = 10)
nfd_obj1 <- Nfd(data = data_matrix1)
nfd_obj2 <- Nfd(data = data_matrix2)
mvnfd_obj <- Mvnfd(nfd_obj1,nfd_obj2)
print(mvnfd_obj)
nfd_obj$data
nfd_obj$nobs

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{A numeric matrix of data points.}

\item{\code{nobs}}{Number of observations.}

\item{\code{nvar}}{number of variables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mvnfd-new}{\code{mvnfd$new()}}
\item \href{#method-mvnfd-print}{\code{mvnfd$print()}}
\item \href{#method-mvnfd-clone}{\code{mvnfd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvnfd-new"></a>}}
\if{latex}{\out{\hypertarget{method-mvnfd-new}{}}}
\subsection{Method \code{new()}}{
Constructor for `mvnfd` objects (same as 'Mvnfd')
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvnfd$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A `nfd` objects which have separated by comma}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvnfd-print"></a>}}
\if{latex}{\out{\hypertarget{method-mvnfd-print}{}}}
\subsection{Method \code{print()}}{
Print method for `mvnfd` objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvnfd$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to be passed to `print`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvnfd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mvnfd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mvnfd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
