% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfdClass.R
\name{nfd}
\alias{nfd}
\title{Define a Set of Multidimensional Non Functional Data Objects}
\usage{
nfd(data)
}
\arguments{
\item{data}{A numeric matrix of data points.}
}
\description{
The `nfd` class represents a set of multidimensional vector data.
Vector data objects are constructed by specifying a matrix of data points.
}
