% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{norm_hd}
\alias{norm_hd}
\title{Compute the norm of an object of class `hd`}
\usage{
norm_hd(hd_obj)
}
\arguments{
\item{hd_obj}{An `hd` object}
}
\value{
The norm vector of the an object of class `hd`
}
\description{
Compute the norm of an object of class `hd`
}
\seealso{
\code{\link{hd}},\code{\link{mvbasismfd}}
}
