% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfdMethods.R
\name{norm_nfd}
\alias{norm_nfd}
\title{Compute the norm objects of object `nfd`}
\usage{
norm_nfd(nfd_obj1)
}
\arguments{
\item{nfd_obj}{An `nfd` object}
}
\value{
The norm of `nfd` objects
}
\description{
Compute the norm objects of object `nfd`
}
\seealso{
\code{\link{nfd}}
}
