% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdMethods.R
\name{norm_vd}
\alias{norm_vd}
\title{Compute the norm of an object of class `vd`}
\usage{
norm_vd(vd_obj)
}
\arguments{
\item{vd_obj}{An object of class `vd`}
}
\value{
The norm vector of the an object of class `vd`
}
\description{
Compute the norm of an object of class `vd`
}
