% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnfdMethods.R
\name{+.mvnfd}
\alias{+.mvnfd}
\title{Addition of two `mvnfd` objects}
\usage{
\method{+}{mvnfd}(obj1, obj2)
}
\arguments{
\item{obj1}{An `mvnfd` object}

\item{obj2}{An `mvnfd` object}
}
\value{
An `mvnfd` object
}
\description{
Addition of two `mvnfd` objects
}
\seealso{
\code{\link{mvnfd}},\code{\link{mvbasismfd}}
}
