% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predictor}
\alias{predictor}
\title{Subject trajectory estimation}
\usage{
predictor(MHPCA, d, functional, regional)
}
\arguments{
\item{MHPCA}{MHPCA object}

\item{d}{group}

\item{functional}{functional dimension}

\item{regional}{regional dimension}
}
\description{
Subject trajectory estimation
}
