% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgMethods.R
\name{print.img}
\alias{print.img}
\title{Title}
\usage{
\method{print}{img}(x, ...)
}
\arguments{
\item{x}{a matrix}

\item{...}{other args}
}
\value{
`img` class
}
\description{
Title
}
\examples{
 x <- matrix(rnorm(100),ncol = 10)
img_obj <- img(x)
}
