% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfdMethods.R
\name{print.nfd}
\alias{print.nfd}
\title{Print Method for `nfd` Objects}
\usage{
\method{print}{nfd}(x, ...)
}
\arguments{
\item{x}{An object of class `nfd`.}

\item{...}{Additional arguments passed to the default print method.}
}
\value{
Invisibly returns the input object `x` after printing.
}
\description{
This function provides a custom print method for objects of class `nfd`.
It displays a header indicating that the object is an `nfd` object
and then proceeds to print the underlying matrix data dimension.
}
\examples{
# Create a standard matrix
mat <- matrix(1:9, nrow = 3, ncol = 3)

# Convert the matrix to an 'nfd' object
nfd_obj <- nfd(mat)

# Print the 'nfd' object
print(nfd_obj)

}
