% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdMethods.R
\name{scale}
\alias{scale}
\title{scale of an object of class `mfd`, `mvmfd`, `nfd`, `mvnfd` or `hd`..}
\usage{
scale(x, ...)
}
\arguments{
\item{x}{An object of class `mfd`, `mvmfd`, `nfd`, `mvnfd` or `hd`.}

\item{...}{all `scale` function arguments.}
}
\value{
An object of class `mfd`, `mvmfd`, `nfd`, `mvnfd` or `hd`.
}
\description{
Standard deviation an object of class `mfd`, `mvmfd`, `nfd`, `mvnfd` or `hd`..
}
