% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnfdMethods.R
\name{scale_mvnfd}
\alias{scale_mvnfd}
\title{Scale an `mvnfd` Object}
\usage{
scale_mvnfd(mvnfd_obj, weight = NULL)
}
\arguments{
\item{mvnfd_obj}{An object of class `mvnfd`.}

\item{weight}{An optional numeric vector of scaling factors for each variable. If NULL, scaling factors are calculated automatically.}
}
\value{
A scaled mvnfd object.
}
\description{
This function scales an `mvnfd` object by calculating a scaling factor based on the variance of its evaluations 
or using a provided weight. It returns a new scaled `mvnfd` object.
}
