% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfdMethods.R
\name{scale_nfd}
\alias{scale_nfd}
\title{Scale a `nfd` Object}
\usage{
scale_nfd(nfd_obj, weight = NULL)
}
\arguments{
\item{nfd_obj}{A `nfd` object to be scaled.}

\item{weight}{An optional numeric value to use as the scaling factor. If NULL, the scaling factor is calculated automatically.}
}
\value{
A scaled `nfd` object.
}
\description{
This function scales a `nfd` object using a provided weight or by calculating a scaling factor 
based on the variance of the object. The scaled `nfd` object is returned.
}
\examples{
# Example usage:
# Assuming `nfd_obj` is a valid numeric functional data object:
# scaled_nfd <- scale_nfd(nfd_obj)
# scaled_nfd <- scale_nfd(nfd_obj, weight = 0.5)
}
