% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{[.hd}
\alias{[.hd}
\title{Extract subsets of an `hd` object}
\usage{
\method{[}{hd}(hd_obj, i = "index", j = "index")
}
\arguments{
\item{hd_obj}{An `hd` object}

\item{i}{An index or indices specifying the subsets to extract for the first dimension}

\item{j}{An index or indices specifying the subsets to extract for the second dimension}
}
\value{
An `hd` object containing the specified subsets
}
\description{
Extract subsets of an `hd` object
}
\seealso{
\code{\link{hd}},\code{\link{mvbasismfd}}
}
