% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnfdMethods.R
\name{[.mvnfd}
\alias{[.mvnfd}
\title{Extract subsets of an `mvnfd` object}
\usage{
\method{[}{mvnfd}(mvnfd_obj, i = NULL, j = NULL)
}
\arguments{
\item{mvnfd_obj}{An `mvnfd` object}

\item{i}{An index or indices specifying the subsets to extract for the first dimension}

\item{j}{An index or indices specifying the subsets to extract for the second dimension}
}
\value{
An `mvnfd` object containing the specified subsets
}
\description{
Extract subsets of an `mvnfd` object
}
\seealso{
\code{\link{mvnfd}},\code{\link{nfd}}
}
