% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdMethods.R
\name{*.hd}
\alias{*.hd}
\title{Multiplication of an `hd` object with a scalar}
\usage{
\method{*}{hd}(obj1, obj2)
}
\arguments{
\item{obj1}{An `hd` object or a scalar}

\item{obj2}{An `hd` object or a scalar}
}
\value{
An `hd` object
}
\description{
Multiplication of an `hd` object with a scalar
}
\seealso{
\code{\link{hd}},\code{\link{mvbasismfd}}
}
