% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnfdMethods.R
\name{*.mvnfd}
\alias{*.mvnfd}
\title{Multiplication of an `mvnfd` object with a scalar}
\usage{
\method{*}{mvnfd}(obj1, obj2)
}
\arguments{
\item{obj1}{An `mvnfd` object or a scalar}

\item{obj2}{An `mvnfd` object or a scalar}
}
\value{
An `mvnfd` object
}
\description{
Multiplication of an `mvnfd` object with a scalar
}
\seealso{
\code{\link{mvnfd}},\code{\link{nfd}}
}
